+++ src/driver/mysql/MysqlDriver.ts
@@ -81,7 +81,7 @@ export class MysqlDriver implements Driver {
     ];
 
     /**
-     * Orm has special columns and we need to know what database column types should be for those types.
+     * ORM has special columns and we need to know what database column types should be for those columns.
      * Column types are driver dependant.
      */
     mappedDataTypes: MappedColumnTypes = {
@@ -184,21 +184,21 @@ export class MysqlDriver implements Driver {
     /**
      * Escapes a column name.
      */
-    escapeColumnName(columnName: string): string {
+    escapeColumn(columnName: string): string {
         return "`" + columnName + "`";
     }
 
     /**
      * Escapes an alias.
      */
-    escapeAliasName(aliasName: string): string {
+    escapeAlias(aliasName: string): string {
         return "`" + aliasName + "`";
     }
 
     /**
      * Escapes a table name.
      */
-    escapeTableName(tableName: string): string {
+    escapeTable(tableName: string): string {
         return "`" + tableName + "`";
     }