+++ src/driver/websql/WebsqlDriver.ts
@@ -165,21 +165,21 @@ export class WebsqlDriver implements Driver {
     /**
      * Escapes a column name.
      */
-    escapeColumnName(columnName: string): string {
+    escapeColumn(columnName: string): string {
         return columnName; // "`" + columnName + "`";
     }
 
     /**
      * Escapes an alias.
      */
-    escapeAliasName(aliasName: string): string {
+    escapeAlias(aliasName: string): string {
         return aliasName; // "`" + aliasName + "`";
     }
 
     /**
      * Escapes a table name.
      */
-    escapeTableName(tableName: string): string {
+    escapeTable(tableName: string): string {
         return tableName; // "`" + tableName + "`";
     }