+++ test/functional/connection/connection.ts
@@ -220,8 +220,9 @@ describe("Connection", () => {
         afterEach(() => closeTestingConnections(connections));
         it("database should be empty after schema sync", () => Promise.all(connections.map(async connection => {
             await connection.syncSchema(true);
-            const queryRunner = await connection.driver.createQueryRunner();
+            const queryRunner = connection.driver.createQueryRunner();
             let schema = await queryRunner.loadTableSchemas(["view"]);
+            await queryRunner.release();
             expect(schema.some(table => table.name === "view")).to.be.false;
         })));
 
@@ -303,8 +304,9 @@ describe("Connection", () => {
                 const commentRepo = connection.getRepository(CommentV1);
                 await commentRepo.save(comment);
 
-                const query = await connection.driver.createQueryRunner();
-                const rows = await query.query(`select * from "${schemaName}"."comment" where id = $1`, [comment.id]);
+                const queryRunner = connection.driver.createQueryRunner();
+                const rows = await queryRunner.query(`select * from "${schemaName}"."comment" where id = $1`, [comment.id]);
+                await queryRunner.release();
                 expect(rows[0]["context"]).to.be.eq(comment.context);
             }));